[network] No error if no IPv6 interface is found
We stop propagating an error if No IPv6 address is found. We just don't
append the IPv6 address to the 'network' map. The payload just contains
IPv4 and and MAC.

This solves the issue an org was having because he doesn't have an IPv6
interface enabled on his CentOS6 machines.