Update python check API to match agent5 signature.
Since we want to import python checks from the agent5 we need to mimic
its api. 'init_config' is kept from agent5 but 'agentConfig' is
deprecated. An empty dict is passed for now since most check await the
parameter but dont use it. We still need to decide how to handle custom
information offered by 'agentConfig' in the agent5.