Avoid sharing internal states between multiple instances of a check.
When a check had multiple instances 'newCheck := c' would not create a
new instance of the underlying struct of the interface. This means that
we end up with multiple reference to the same struct and the
informations of the last instance would apply to all. One the the side
effect would be that all check of a certain type would share the same
Sender (which makes Rates failed).

To avoid that we now register a factory function that allow us to
instantiate a new struct of the check type.