[collector][aggregator] Implement Check IDs (#138)
Each check implements the `ID` function, which can use a helper
that generates the ID from the check name, instance config and
init config. With the helper, the instance and init config are not
generated from the semantic config but from the raw byte slice.

The ID is used by the aggregator too.