Fix CPU check erroneous Rates.
The CPU check use to send percentage as rates. This produce a total
percentage lower than 100%. This is due to the use of time and not
number of cycles in the aggregator (we always had some variation in time
range).

The CPU check now saves last CPU information as an internal state and
produce percentage by hand as Gauge.