Adding a method to acces agent configuration from python check.
The 'config_get' method aim to replace the 'agentConfig' parameter from
the python checks.

We add the ability to convert simple types (int, string, ...) from Go to
Python. Some Go types can't be converted to Python for now, we will had
more types if needed in the futur, for now we emit a simple warning.