[aggregator] Add support for events submission from checks
Overall, events are handled in a similar way as service checks.

The whole event struct has to be submitted from the agent checks, this
allows handling all the optional fields of the event struct more
nicely (specifying them all as parameters of an `Sender.Event` function
wouldn't be very handy from checks).

The "Agent startup" event is slightly tricky to handle so it's not
taken care of here.

The events are sent to the agent 5 intake endpoint (see comments for
explanation).

For now:
- Haven't added the python bindings for py checks yet
- It only supports the v1 api endpoint (some work is queued to support
the v2 endpoint)