[python] Add bindings to submit events from py checks
The go code is written to be safe, but some edge cases may still not be
handled properly.

The approach to convert the python dict to a go Event struct is not
generic at all but I didn't find a strong reason to make it generic.

There is some (limited) duplicated logic with some of `go-python`'s
internal api.

Known limitation: doesn't support `unicode` fields in event dict (yet).