[python] Bubble up error correctly when passed `tags` is not a sequence
This is one way of dealing with a CPython call returning an error.
Another way would be clear the error from the python interpreter with
`PyErr_Clear()` and handle it entirely from the go/C code.

In any case we should always handle errors from CPython correctly
otherwise the python interpreter can raise the exceptions later in the
code, which is extremely confusing.