[python] Add bindings to submit events from py checks (#207)
The go code is written to be safe, but some edge cases may still not be
handled properly.

The approach to convert the python dict to a go Event struct is not
generic at all but I didn't find a strong reason to make it generic.
There is also some (limited) duplicated logic with some of `go-python`'s
internal api.

We encode unicode objects to python strings with `utf-8` encoding so
that go can handle them easily downstream. Catch and log any error.
The implicit assumption here is also that any python string is encoded
with `utf-8`.

Also, we bubble up errors correctly when passed `tags` is not a
sequence. This is one way of dealing with a CPython call returning an error.
Another way would be to clear the error from the python interpreter with
`PyErr_Clear()` and handle it entirely from the go/C code.