[py] Log complete python errors + tracebacks when check import fails
In some cases the python module of a check can't be imported because
of a python code issue, in that case logging the traceback is useful.

Use a common function that retrieves and formats the python
interpreter error, and clears the error flag in the interpreter.
Any python c api call that returns an error should be followed by
a call of that common function to get a formatted string of the error.