[py] Don't use a nested StickyLock
Using a nested lock unlocks to goroutine to the thread when the child
lock is unlocked (whereas we'd want that to happen when the parent lock
is unlocked).

Until we can figure out a good way to deal with this situation, don't
use nested locks, and add a warning to `getPythonError`.