[agent build] Set `RPATH` with `-r` ldflag in rake build task (#256)
And remove the custom `LD_RUN_PATH` from the environment when the agent
is built with omnibus.

This allows creating the same `agent.bin` binary whether the agent is
build with omnibus or with the `agent:build` rake task without omnibus.

Without this change, building an agent locally with `agent:build`
would create a binary without the `RPATH` set to the embedded lib dir,
so it wouldn't find the libraries located there. This would in turn
make the embedded python interpreter unable to find some libraries.


Also, exit if `go build` is unsuccessful.