[kubernetes] Add metadata provider
Based on the data we are currently collecting in the `kubeutil` module
in `dd-agent`, this returns a payload format that lists all active
deployments, replica sets, services, pods and containers.

All data types can be "joined" via IDs in each model so we can resolve a
single container id with all the relevant metadata or treat each bit
independently for visualization, etc.

The goal would be to have this run on a subset of Agents which would
target the master and send this metadata to our systems for backend
resolution.