[py] Use same stickyLock when fetching the interpreter error
Changing locks between a python call that sets the error indicator
on the interpreter and a fetch of that error can result in the
error indicator not being there anymore when the fetch is attempted.

I'm not sure exactly why this happens. The fact that it happens
"randomly" makes me think that it's related to how the goroutines are
scheduled on OS threads by the go runtime, and to how this can affect
the state of the python interpreter.