Make text for cfgpath flag clearer  (#289)
If you set the path in the following manner: ./bin/agent/agent start -f $(pwd)/datadog.yaml, it will use the one in ./bin/agent/dist/. This is because viper does not see a path and think, that's where the file is. It sees a path and thinks, that is a directory the file must be in. (See the logic https://github.com/spf13/viper/blob/0967fc9aceab2ce9da34061253ac10fb99bba5b2/viper.go#L1533-L1559.)

This updates the text to make it more clear that it should be directed at the directory the config resides in, rather than at the config file itself.

(We could also do some logic to add the parent path rather than the file itself, however I think this will result in less misunderstanding, as it has to be named datadog.yaml.)