[py] Add `increment`/`decrement` methods to `AgentCheck`
Eases the transition from agent 5 to agent 6 for the checks. The
methods use `count` after adding a suffix to the metric name.

The suffix is required because `count` submits metrics with the
`COUNT` api metric type, whereas in agent 5 these methods use the
`RATE` api metric type, so we need to use a different metric name
(the backend doesn't support using a different API metric type for
the same metric name).