[py] Improve handling of py interpreter errors (#290)
* Fix error handling by using the same `stickyLock` when fetching the
interpreter error

Changing locks between a python call that sets the error indicator
on the interpreter and a fetch of that error can result in the
error indicator not being there anymore when the fetch is attempted.

I'm not sure exactly why this happens. The fact that it happens
"randomly" makes me think that it's related to how the goroutines are
scheduled on OS threads by the go runtime, and to how this can affect
the state of the python interpreter.

* Various improvements on `getPythonError`:
  * now a method of the `stickyLock` struct for clarity
  * check that an error occurred before trying to fetch it (improves
  quality of error that's returned by the function)
  * normalize the python error (recommended by the python C API docs)
  * use the string representation of `pvalue` (improves error message)


* Use `getPythonError` in py's `check.Run` to fetch errors from the
python interpreter. We should use it wherever possible.