[aggregator] Add `source_type_name` to metric series
This field is important for backwards-compatibility on the backend:
different `source_type_name`s lead to different contexts.

By default all metrics posted to the metrics API get assigned the
`API` source type name, but metrics coming from the agent's collector
have always had the `System` source type name.

In order to use the same contexts as the agent5, we need to set this
source type name on all metrics coming from the collector.

This is unlikely to change in the backend in the near future.

The changes here only apply to the v1 API, another commit will take
care of the v2 API.