[aggregator] Use `device` field for v1 metric endpoints
Unfortunately the v1 metric endpoint does not currently handle a
`device:` tag in the same way as the `device` field. To avoid a
breaking change on this during an upgrade from agent5 to 6, use the
`device` field for the v1 metric endpoint.

Given that the v2 endpoint should handle `device:` tags correctly,
I use a quick fix and populate the `device` field right before
the serialization for the v1 endpoint. I left the serialization
itself untouched as it's quite error-prone.