[forwarder] Fix api key for multiple v1 endpoints
v1 endpoints wouldn't work with multiple endpoints because all the
endpoints would use the same api key, set by the aggregator.

The forwarder should be the only piece of the agent that's aware of the
api keys of each endpoint, and should decide when to use them.

Remove all usage of `api_key` outside of the forwarder, except when
used as a "placeholder" api key that are needed in some legacy payloads