[forwarder] Drop transaction when reponse is 404
Like the 400 and 413 http codes, 404 means that the backend is healthy
but the request is invalid. It makes sense to drop the transaction in
that case.

Useful in particular for endpoints that are not implemented on all the
accounts/environments, and agents with multiple domains/keys
configured. Dropping the 404 transactions allows not filling up the
retry queue with these transactions.