Move serialization to metrics and let the serializer route payload to the forwarder
The serialization has been moved to the metrics so their internal logic it
kept there.

The serializer has been improved to handle the routing of metrics to the
forwarder. Different parts of the agent don't have to be aware of which
endpoint or format to use. This will allow us a dynamic routing (which
format/endpoint per metric type) and simplify the flow of metrics
through the agent.