[collector/loader] Fix nil dereference when a loader fails initializing (#401)
Since each loader (as returned by LoaderCatalog()) is an interface,
it's not nil when the underlying concrete loader failed
initializing (and is actually nil). Once the loaders are added to
AutoConfig, we assume they're not nil, which leads to nil dereferences.

To fix this, make the factory functions return an error when a loader
initialization fails, and skip those loaders.