resolves a data race in autoconfig (#414)
Logging does not happen inline, it is handed off to a background thread. This is fine, as long as you only pass it objects that are threadsafe. This object is not, we have to use the mutex to ensure safety. Unfortunately, that means that when the log hands it off to another thread to log the line, it creates a data race. Formatting it inline will ensure that the object is not unsafely handed off to another thread.