[serializer] Add config param to use v2 api (#441)
* Add config option to switch to v2 endpoint

Per endpoint config option.

For now we use the v1 endpoints by default, until the v2 endpoints are
considered stable.

* Add `Content-Encoding` header when compression is used

The intake uses that header to determine if/how to decompress the
payloads, so let's add it in the serializer.

* [serializer] Disable compression on v1 series/check_run endpoints

zstd compression is supposed to work on the v1 series and check_run
endpoints, but for some reason it doesn't. Let's disable it until we
have time to investigate more.