[py] Cast metric values to floats
The python bindings can already handle casting `int`,
`decimal.Decimal`, etc, to floats, but some other types currently
throw exceptions, for example `string`s (and yes, some checks
do send numeric values as strings).

To address this, explicitly cast values to floats in the `AgentCheck`
class. This will still (rightfully) raise exceptions when the value
is not castable to a float, but it should make the behavior of
Agent6 closer to Agent5's.