[py] Implement backward-compatible `increment`/`decrement`
Some experiments on existing checks have shown that sending these
metrics with `count` was not working well in some of the checks
(namely those that should be using `gauge` instead of `increment`/
`decrement`).

Until we find a reasonable deprecation path to remove these methods,
let's re-implement them fully for python checks (but keep the
deprecation warnings).