Move 'ac_exclude' and 'ac_include' containers list to datadog.yaml
This logic impact the docker check but also autoconfig, the tagger,
orchestrators... For consistency it must be part of the agent internal
and not just the docker check.

This also requires a common initialization of dockerUtil to have the
same behaviour everywhere. The exposed API from util/docker is in charge
of returning null value if not initialized.