[docker] Update network detection logic to handle different modes.
Check the NetworkMode in the HostConfig first to determine if a
container is in host mode, no-network mode or depends on a separate
container id to provide the network namespace.

If we're in the last case we will also resolve these cases to their
respective network namespace after we've gathered all container
networks.

Tests are added around all of these cases.