[collector] Increase default number of check runners
We need to increase the default number of workers now that we can
expect a lot of agents to run 4 long-running checks (JMX, apm, process,
and logs), otherwise by default no runners would be left for the
"regular" checks.

Increase it to 20, since some testing/benchmarking has shown that 20
runners is a good compromise for most setups.