Fix the 'status' command. (#727)
Change `GetDistPath()` to return the path relative to the binary when
the registry entry isn't found.  The registry entry isn't strictly
required, and this will make the code more resilient.

Change the status command to load the `dist` path (to find the
templates) using the `common` method `GetDistPath()`, rather than
computing itself, so that it will function in a platform neutral way.