temporary fix for tagger cache misses
We want to store the empty tag response if the error
comes from successfully parsing the source but not
finding the entity (can happen in k8s/ECS in pure
docker containers).
On other cases, don't store that to retry on next query.

Implemented for now with string matching, need to create a custom
error type to handle that