[packaging] Fix upstart job definition for RHEL 6 (#808)
`console log` is not supported by the version of upstart that RHEL 6
provides. Instead, we have to redirect the process' output to a file
and manually rotate it in a pre-start script.

This also divides the upstart jobs into 2 (one for the deb package, the
other for the rpm package): it makes more sense given the large
differences in features between the RHEL 6 upstart and the recent
Debian/Ubuntu upstart, and to keep the Deb/Ubuntu one as simple as
possible.