[omnibus] Unpin the external agents and integrations-core
On master, we should build with the latest versions of these deps,
otherwise it's hard to synchronize changes that affect both
`datadog-agent` and these other projects.

For now, we can pin them in the beta build by pinning the deps on the
`beta` branch. If at some point we want to build beta/stable builds
from master we'll have to discuss how to handle that.