[gui] Write token to run directory
The agent user doesn't necessarily have write access to the conf
directory, so writing the auth token there can fail.

Let's write the token to the `run` path, which is a path where
(by definition) the agent user should have write access. The read
permissions on this directory should be as limited as the conf
directory.