[check cmd] Fix discrepancy in stats displayed by check cmd (#977)
When the check `Commit`s its metrics/service checks/events, in the
sender we should simply "rotate" the stats structs sequentially so we
know they're up-to-date as soon as the `Commit` is done. I don't see
any benefit of doing that in a separate goroutine.

Also, rename a function and make a comment more accurate.