[jmx] Remove conf fields from instance struct (#976)
These fields aren't used anywhere and can cause type mismatches between
what's in the conf instances and the types defined in the struct
fields, even when downstream JMXFetch can accommodate different types
for some fields. (These mismatches make the jmx check throw errors)

If we want to introduce more agent-side validation of the jmx configs
we should do it directly in JMXFetch (and it already does a bit).