remove resversion handling from podwatcher
Comparing resversions accross namespaces is not supported
(might be sharded), so we remove that logic.

The new container logic is enough for our uses, as all users
drop pods with no containers for now. We might need it back if
we want tagger to collect `pod://` tags in the future