[install_script][centos] Fix failures on repo/package install (#1145)
* [install_script] Use a unique repo name on RHEL/CentOS and SUSE

The repo name needs to be unique on RHEL/CentOS, otherwise the user gets
this error: `Repository datadog is listed more than once in the
configuration`, and inconsistent behavior.

If the user is upgrading from Agent 5, the `[datadog]` name is used by
the stable repo, so use `[datadog-beta]` instead (it's reflected in the
file name too).

Do the same thing on SUSE (not sure it's required there, but it makes
sense anyway).

Update docs too.

* Remove `yum check-update` from install script

This command should be used to checked if updates of existing packages
are available, and fails if updates are indeed available, which makes
the whole script fail.

That's not the behavior we want. Let's remove it. `yum install` will
download the missing or expired metadata of all repos.