Fix handling of empty fields in `resources` and `host` metadata payloads (#1286)
* [resources] Remove the 'resources' key from the payload if it's empty

As crazy as it sounds, sobotka can crash if the 'resources' key is
present in the payload but is empty/isn't fully populated.

Let's just make sure we don't have the key in the payload if there's
no data in it.

* [resources] Only enable resources collector on Linux

To reflect A5 more closely, where sending resources regularly is
only done on Linux by default (on gohai's side it's not implemented
on Windows and can cause some CPU spikes on macOS).

* [host metadata] Omit empty `host-tags` entries in JSON payload

The backend expects an array under every key in `host-tags`. Let's just
omit any empty entry when JSON-serializing.