[Logs] Enable protobuf transport by default
Previously logs would be submitted as strings, following the rfc5424
protoco. In parallel, we had an option to submit logs using protobuf.
This commit defaults the transport to protobuf.
Note, we have to use a different endpoint. A different port would impact
servers that have whitelisted outbound ports, so instead we opted for
a different dns.
It is important to note that `intake-agent.logs.datadoghq.com` on port 10516
will accept protobuf only. Thus, to disable this we would need to both
set `dev_mode_use_proto` to false, and update the `dd_url`.
We're fine with that, as this should be transparent for the user, and this
option was added only for testing purposes. A later commit will remove the
option, and all agents will use protobuf.