[forwarder] Increase default recovery interval from 1 to 2
Once an endpoint starts recovering, let's be a bit more aggressive in
the forwarder's recovery speed to avoid hitting the max retry queue
size when the endpoint is available again.

A value of 2 doesn't avoid that completely but at least mitigates a bit
the increase of the retry queue size, while probably being
"nice enough" for the backend.