[reno] Configure regexes to match our release and pre-release tags
Currently, reno doesn't properly understand which tags are release tags
and which are pre-release tags because its default regexes don't match
our tagging conventions. This makes release notes hard to generate
properly for a release when RC tags are present.

This adds a configuration file to reno with regexes that match our
tagging convention.