[metadata] Fix collection of project GCP host tags
The `project` and `numeric_project_id` tags weren't collected because
they weren't pulled from the correct key of the JSON response. This was
a regression from Agent 5.

We actually need to fetch the whole metadata JSON instead of just
the `instance` key to have access to the project metadata.

Fixes this, and updates the test fixture to reflect what the API
actually returns.