[packaging][linux] Create dd-agent user without a login shell (#1800)
i.e. set `nologin` as the login shell of the `dd-agent` user when
creating it. This follows most distros' recommendations when creating
system users such as `dd-agent`.

If the `dd-agent` user already exists, do not change its login shell to
allow users to change it if they want and persist that across upgrades.

The path to `nologin` is different between RHEL/SUSE and debian/ubuntu,
account for that.