[packaging] Avoid install failure if enabling systemd service fails
When `systemctl enable datadog-agent` fails, the package install
currently fails. The root cause is that our test for systemd,
`command -v systemctl`, tests for the presence of the systemctl
command but not that systemd is used as the init system on the current
system. This can happen for example on Ubuntu 14.04 (-> upstart) when
the `systemd` deb is installed.

Unfortunately it seems very hard to reliably determine what the current
init system is across all distros (even the cmdline of PID 1 isn't
always enough). The real fix would be to ship a different package for
every distro/version tuple, with pkg scripts tailored for that
distro/version.

So for now, let's just ignore these errors. Worst case scenario, on the
few systems where both commands exist, the agent install may not
stop/restart the agent service properly.