Allow omitting GCE metadata from host tags
Similar logic already exists for EC2: When collect_ec2_tags is set to
false, no extra tags are generated based on EC2 metadata. This unifies
the behavior so that GCE tags can also be excluded (though default
value for the new setting is different from the existing EC2 setting to
avoid changing existing behavior).

Motivation for the change is that GCE metadata may contain confidential
information (as it does in our case) and that must not be sent outside
the execution environment.

This also allows making the agent work like it did with version 5: With
version 5, if metadata collection was off but custom host tags were
specified, the tags got associated with the metrics. With version 6 the
tags are only ever sent when metadata collection is enabled, which also
sends the undesired GCE tags.