[test][runner] Fix flaky test
This should fix what I think is the root cause of the flakiness:
the 2 checks had the same ID, so with multiple runners, if the check c2
check was run before check c1, c1 could be discarded because c2 (with
the same ID was already running).

To fix this, give them different IDs.