Allow checks config to fail independently of logs config (#2432)
### What does this PR do?

Allows the logs config to be sent if the checks config parsing failed in AD. 
I refactor the `extractTemplatesFromMap` function to return an array of errors. I add the info whether checks or/and logs parsing failed.

I still handle the errors in the `providers` to give context on the errors (did it happen in a pod / docker container / ecs ?)

### Motivation

In Autodiscovery, when the parsing of checks config fails, the logs configs is not parsed and fails. We want to be able to send logs even if the checks config is broken

### Additional Notes

Ideally we would display something in the `agent status` (or `agent configcheck`?) when we fail to parse a config. I'm not sure there is a simple way of doing this for the logs.