[windows] Fixes crash on malformed windows counter database.
We sometimes encounter malformed windows performance counter databases.
We try to handle it gracefully.  However, we recently encountered a
database which was filled with empty strings.

Code change ensures we don't try to access off the end of the end of the
array of strings, which was possible in this particular case