Add "Tolerate Unready" Annotation
This will bypass the normal logic that skips pods that aren't ready.
It's a bit of a divergence from the normal way of doing annotations,
which is that they're based on the container.

This status is global to the pod, though, so I had to come up with
something.
If you were to have a container named `tolerate-unready` this would be
confusing, but still wouldn't collide because it'd still be looking for
things suffixed off that (`.checks` or whatever)

I had a lot of trouble with the testing framework, so I simply adapted
one of the tests I could convince to run on my system to require these
changes by making one of the fixture pods unready, but adding the
annotation so the tests still pass.