pkg/config/legacy: enable API to support additional legacy keys from callers
This change adds a new method (`Load`) to the `legacy` package allowing
callers to set further sections and keys from the Agent 5 configuration
onto the new configuration by means of a callback.

This is to enable agents (such as process and trace) to continue adding
new entries to the old configuration without needing to modify core
code.