[flare] refactor using a state during the archive creation instead of writing each time to the file.
In order to not open the permissions.log file for every lines. I've used
a new struct to store filepaths that will be added to the
`permissions.log` file. Before finally creating the .zip archive, we create
the `permissions.log` file with all files information in one shot.

Slightly change the signature of two methods shared with the DCA:
  - `zipConfigFiles`
  - `zipLogFiles`

This commit doesn't add the `permissions.log` file for the DCA flare archive,
this is why the nil pointer is handled in those methods (to avoid the DCA to be
forced to provide a *PermsInfos instance).