[flare] add a permissions.log file into the flare archive (#2707)
* [flare] implement methods to add permissions info of files added to the flare archive.

Following the format of the permissions.log generated in the Agent 5
except for the mode info which looks like an `ls -l` call:
`-rw-r--r--` instead of `0644` in the old format.

* [flare] creates the permissions.log file in flare while creating the archive.

Only for non-windows system, windows generated flares still won't have
any `permissions.log` file.

* [flare] unit tests for the permissions.log file writing.

* [api] expose the auth token filepath.

* [flare] add auth token permissions infos.

* Add release note entry.

* [flare] fix perms file method signature for windows build.

* [flare] refactor using a state during the archive creation instead of writing each time to the file.

In order to not open the permissions.log file for every lines. I've used
a new struct to store filepaths that will be added to the
`permissions.log` file. Before finally creating the .zip archive, we create
the `permissions.log` file with all files information in one shot.

Slightly change the signature of two methods shared with the DCA:
  - `zipConfigFiles`
  - `zipLogFiles`

This commit doesn't add the `permissions.log` file for the DCA flare archive,
this is why the nil pointer is handled in those methods (to avoid the DCA to be
forced to provide a *PermsInfos instance).

* [flare] fix a bug about reading the copied file instead of the original one.

* [flare] add a missing err test while writing permissions.log header

* [flare] PermsInfos func should not be public.

We don't need them from outside of this pkg.

* [flare] fix permissions check for log files (was using the dst file).

* [flare] changed the PermsInfos composition and made it private.

Bonus, `permsInfo` becomes `filePermsInfo`.

* [flare] archive dca flare generation also create a `permissions.log` file in the archive.

* [flare] wrong func assignement for windows stubbed funcs.

* [flare] while creating dca flare archive, do not fail on permissions.log creation error, just log the error.