[windows] fix downgrade handling. (#2789)
* [windows] set the "wait" option on the windows services to yes, so that
those operations are blocking

* [windows] fix downgrade.

Not sure how *upgrade* worked, as they're treated the same.  Root cause
appears to be that we were adding the subservices twice, and so
reference counting was thrown off.  This is made worse by the fact
that the omnibus processing explicitly ignores that error code.

Also added a subdirectory for recreating the MSI w/o an entire omnibus
build, which cuts a lot off the build/test cycle of installs.

* Add documentation for local build script