[check cmd] Add pause between 2 consecutive runs with `--check-rate` (#2793)
* [check cmd] Add pause between 2 consecutive runs with `--check-rate`

`--check-rate` is meant to be used for the metrics that need the
check to run twice before reporting all their metrics. Some of these
checks need some time to pass between their 2 runs, otherwise the runs
are too close in time to provide meaningful samples. This affects the
go cpu check for example.

This sets a pause of 1 seconds between the 2 check runs when
`--check-rate` is used. The option can be set separately with `--pause`
if `--check-times` is used instead (default: no pause).