[omnibus] Remove protobuf-py from omnibus deps of Agent
protobuf-py is needed by integrations-core. On Linux, we want
to ship protobuf-py with the cpp extension for perf reasons, and these
were not available as binary wheels with the `ucs2` build of the python
embedded Agent. This explains why we originally built the protobuf lib
in omnibus.

We now build python with `ucs4`. protobuf-py+cpp binary wheels are
provided for this build. These wheels use the `manylinux1` policy,
which support CentOS 5.
So we're all good relying on these binary wheels instead.