pkg/trace/api: never drop traces
This change modifies the HTTP receiver to never drop traces. Dropping
traces does not bring any real benefit. If a user wishes to process a
high amount of traces, he should be able to do so at the cost of
hardware resources. For now, we are safeguarded against excessive use of
memory and cpu by the watchdog system.