[macOS/apm] Do not disable APM when the old conf option is not set
Likely because of an oversight, the `apm` check that runs trace-agent
on macOS was still checking that the old `apm_enabled` option was set
to `true` before starting trace-agent. This logic would also make
trace-agent not start by default on macOS.

Since at least 6.0.0, trace-agent should be enabled/disabled with the
`apm_config.enabled` setting (or the `DD_APM_ENABLED` env var). This
change also makes the trace-agent start by default on macOS.