pkg/trace/config: turn log throttling on by default (#3309)
This change makes sure that the (informerly deprecated) option
`apm_config.log_throttling` is turned on by default. This means that
error logs are rate limited to a maximum of 10 every 10 seconds
(hardcoded).

Having this always off resulted in countless spammy logs and support
cases. It is useless to address more than 10 error messages at a time,
which are most of the cases the same one.

For those rare cases, a change has been made where log throttling is
automatically disabled when logging level is set to "debug" and log
throttling is not specifically set by the user in the config.