Avoid deadlock when go switch thread
Locking the GIL in the setup could end up in a python deadlock. As go
might switch threads python will end up with a GIL locked but in a
different thread. This would cause 'PyRun_SimpleString' to block
indefinitely.

We now lock the OS thread and acquire the GIL just when we need it (like
we do in the agent).