pkg/trace/config: allow stats payload sends to be parallel
This change allows 10% of the maximum connections (by default 200) to be
used a concurrent connections to the Datadog API when flushing stats.

We've already made this change for traces, and we should make sure we
have it especially for stats, where it is quite common to flush multiple
payloads at once which would normally result in long blocking times in
the pipeline.