[pkg/gui/checks] Prevent crash when creating custom check config with the GUI
The saving of a config file for a custom check created with the GUI could fail
because the directory where the config file should be saved may not exist.
Changes the POST /checks/setConfig endpoint so that the directory where the
custom check config will reside is indeed present before saving the file.