[pkg/config] Fix jmx_check_interval default
The default check interval was not cast to an int before being saved
as the default config. As such, the default JMX check interval was
wrong for two reasons:
- it's a time.Duration of value 15000 (which means 15 microseconds,
and not 15 seconds as expected)
- since it's a time.Duration, when the Windows config file is
written, it's written as 15us (with the micro unicode symbol instead
of the u), but parsed afterwards as an int, which breaks the checks.
To fix these two problems, the default check interval is now cast to
an int.