[pkg/forwarder] Set User-Agent on api key validation request
Pull request #2375 sets the User-Agent string for all forwarder
requests except for the initial /api/v1/validate request which
checks the validity of the api key provided by the user.
This sets the User-Agent to datadog-agent/<agent-version> for the
api key validation request, so that it is the same as all other
requests.