trace/api: add missing expvar handler and fix silent test fail (#3747)
* trace/api: add missing expvar handler and fix silent test fail

Changes:
* restore missing `/debug/vars` endpoint and add corresponding test
* fix silent test failures in `pkg/trace/api/...`

`/debug/vars` was accidentally removed when we switched to a custom mux handler in https://github.com/DataDog/datadog-agent/pull/3308. We partially restored the lost endpoints in https://github.com/DataDog/datadog-agent/pull/3365. This restores another one.

Added an api test to confirm the existence of the expvar endpoint. It now fails if it's not there:

```
--- FAIL: TestExpvar (0.00s)
    api_test.go:764:
        	Error Trace:	api_test.go:764
        	Error:      	Not equal:
        	            	expected: 404
        	            	actual  : 200
        	Test:       	TestExpvar
        	Messages:   	failed to read expvars from local server
```

In order for the test to work `trace/api/api_test.go:TestMain` had to be fixed to call `os.Exit(m.Run())` instead of `m.Run()`. When calling only `m.Run()` it causes all tests in `pkg/trace/api/...` to fail silently. As per [the docs](https://golang.org/pkg/testing/#hdr-Main), *It should then call os.Exit with the result of m.Run*.

We're seeing this problem in the latest RC that we're running. When running `datadog-agent status` it shows an error with the expvar check:

```
      Instance ID: go_expvar:f77c6c92ba32d2ef [ERROR]
      Total Runs: 59,076
      Metric Samples: Last Run: 0, Total: 0
      Events: Last Run: 0, Total: 0
      Service Checks: Last Run: 0, Total: 0
      Average Execution Time : 4ms
      Error: HTTPConnectionPool(host='localhost', port=7654): Max retries exceeded with url: /debug/vars (Caused by NewConnectionError('<urllib3.connection.HTTPConnection object at 0x7fda1a0e9c50>: Failed to establish a new connection: [Errno 111] Connection refused',))
```

* fix linter error

* fix vet resp check order

* s/HandleFunc/Handle

* add release notes

* add APM prefix

* Remove release note.