[gitlab-ci] Move kitchen cleanup phase to last
When kitchen tests do not crash, the VMs used for the tests are
scheduled to be deleted. In this case, the cleanup phase only
waits for the deletion to be done. Instead of waiting 5 to 10
minutes to delete them, this stage has been placed last so that
we waste less time doing nothing.